Fix two bugs
First bug: When there was a negative value on the display and the user
deleted all the numbers, NaN was displayed, because the algorithm tried
to convert "-" to a number. Now, the display value returns to the default.

Second bug: When the user completed an operation with the equal button
and clicked on an operator button, the first number of this new operation
was not the result of the latter operation, but the first number. Now,
it uses the result of the latter operation as the first number of the new one.