Fix a bug
When you made a division by 0 with the "equals" button, there was the error
value on the display, but when you made it using another operator button,
there was only the default value on the display. Now, a division by 0
using either the equals button or any other operator displays the error value.